requirements.txt
pyyaml==6.0.2
jsonschema==4.23.0
jinja2==3.1.5
pytest==8.3.2
